rename -fn to fn; update paths
