+++ lib/api/init.js
@@ -2,7 +2,7 @@
 
 const Promise = require('bluebird');
 const co = Promise.coroutine;
-const $ = require('../-fn');
+const $ = require('../fn');
 const READY = '_ready';
 
 module.exports = co(function * () {
