+++ lib/fn.js
@@ -33,20 +33,7 @@ const $ = {
 			map[el] = el;
 		}
 		return map;
-	},
-
-	shallowCopy: (merge, value) => {
-		/* use object.assign */
-
-		// var copy = merge || {}
-		// Object.keys(value).forEach(function (key) {
-		// 	copy[key] = value[key]
-		// })
-
-		// return copy
 	}
-
-	// diff: require('./diff')
 };
 
 module.exports = $;
