+++ lib/reporter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const print = require('./-fn').formatTime;
+const print = require('./fn').formatTime;
 const fmt = require('./fmt');
 
 const evts = {
